# Optimize indexing of struct fields