@dsnet 2025-01-26
A cool trick for looking for optimizations is to do `go build -gcflags=-S` and look for calls to `runtime.panicSlice`, this is suggestive of the Go compiler not being able to prove that a bounds check could be eliminated. Some massaging of the code might be able to eliminate these checks.

As a bonus benefit, you can reduce binary size as well.